Tim's Demo Game (with fixes)
I am commiting Tim's Demo game so that everyone can access the same thing (I recommend getting the GitHub desktop app). I have fixed the jitter bug and made it so enemies and the player can not push each other. I ran into some problems with the player being pushed out of bounds. This was due to the RigidBody setting 'collision detection' being set to continuous. Changing this to discrete has seemed to fix the issue. I have added a gitignore to this commit to get rid of all the meta files. Let me know if this has broken anything.