Added Enemy Spawning on timer. Fixed enemy attack time bug
I have added in the enemy spawn functionality which is currently working on a timer. So an enemy will spawn once when the timer reaches zero and the timer is also reset so this happens continuosly.
I also fixed the bug that allowed the player to touch the enemy without taking damage by putting the timer countdown outside of the "in attack range" statement as well as having it inside.